Move review components to sub folder
